+++ src/TMModels/Game.cs
@@ -3,11 +3,11 @@
 namespace TMModels;
 
 public record Game(
-    [property: JsonPropertyName("id")] uint Id,
+    [property: JsonPropertyName("id")] int Id,
     [property: JsonPropertyName("name")] string Name,
     [property: JsonPropertyName("isFinished")] bool IsFinished,
     [property: JsonPropertyName("isStalling")] bool IsStalling,
-    [property: JsonPropertyName("turn")] uint Turn,
+    [property: JsonPropertyName("turn")] int Turn,
     [property: JsonPropertyName("map")] Map Map,
     [property: JsonPropertyName("houses")] HouseScore[] Houses,
     [property: JsonPropertyName("winner")] string? Winner,
@@ -16,17 +16,17 @@ public record Game(
 public record HouseScore(
     [property: JsonPropertyName("name")] House House,
     [property: JsonPropertyName("player")] string Player,
-    [property: JsonPropertyName("throne")] ushort Throne,
-    [property: JsonPropertyName("fiefdoms")] ushort Fiefdoms,
-    [property: JsonPropertyName("kingsCourt")] ushort KingsCourt,
-    [property: JsonPropertyName("supplies")] ushort Supplies,
-    [property: JsonPropertyName("powerTokens")] ushort PowerTokens,
-    [property: JsonPropertyName("strongholds")] ushort Strongholds,
-    [property: JsonPropertyName("castles")] ushort Castles,
-    [property: JsonPropertyName("cla")] ushort Cla,
+    [property: JsonPropertyName("throne")] int Throne,
+    [property: JsonPropertyName("fiefdoms")] int Fiefdoms,
+    [property: JsonPropertyName("kingsCourt")] int KingsCourt,
+    [property: JsonPropertyName("supplies")] int Supplies,
+    [property: JsonPropertyName("powerTokens")] int PowerTokens,
+    [property: JsonPropertyName("strongholds")] int Strongholds,
+    [property: JsonPropertyName("castles")] int Castles,
+    [property: JsonPropertyName("cla")] int Cla,
     [property: JsonPropertyName("minutesPerMove")] double MinutesPerMove,
-    [property: JsonPropertyName("moves")] ushort Moves,
-    [property: JsonPropertyName("battlesInTurn")] ushort[] BattlesInTurn) : IComparable<HouseScore>
+    [property: JsonPropertyName("moves")] int Moves,
+    [property: JsonPropertyName("battlesInTurn")] int[] BattlesInTurn) : IComparable<HouseScore>
 {
     public int CompareTo(HouseScore? otherHouse)
     {
@@ -53,26 +53,26 @@ public record Map(
 
 public record Land(
     [property: JsonPropertyName("isEnabled")] bool IsEnabled,
-    [property: JsonPropertyName("id")] ushort Id,
+    [property: JsonPropertyName("id")] int Id,
     [property: JsonPropertyName("name")] string Name,
     [property: JsonPropertyName("house")] House House,
-    [property: JsonPropertyName("footmen")] ushort Footmen,
-    [property: JsonPropertyName("knights")] ushort Knights,
-    [property: JsonPropertyName("siegeEngines")] ushort SiegeEngines,
-    [property: JsonPropertyName("tokens")] ushort Tokens,
-    [property: JsonPropertyName("mobilizationPoints")] ushort MobilizationPoints,
-    [property: JsonPropertyName("supplies")] ushort Supplies,
-    [property: JsonPropertyName("crowns")] ushort Crowns);
+    [property: JsonPropertyName("footmen")] int Footmen,
+    [property: JsonPropertyName("knights")] int Knights,
+    [property: JsonPropertyName("siegeEngines")] int SiegeEngines,
+    [property: JsonPropertyName("tokens")] int Tokens,
+    [property: JsonPropertyName("mobilizationPoints")] int MobilizationPoints,
+    [property: JsonPropertyName("supplies")] int Supplies,
+    [property: JsonPropertyName("crowns")] int Crowns);
 
 public record Sea(
     [property: JsonPropertyName("isEnabled")] bool IsEnabled,
-    [property: JsonPropertyName("id")] ushort Id,
+    [property: JsonPropertyName("id")] int Id,
     [property: JsonPropertyName("name")] string Name,
     [property: JsonPropertyName("house")] House House,
-    [property: JsonPropertyName("ships")] ushort Ships);
+    [property: JsonPropertyName("ships")] int Ships);
 
 public record Port(
     [property: JsonPropertyName("isEnabled")] bool IsEnabled,
-    [property: JsonPropertyName("id")] ushort Id,
+    [property: JsonPropertyName("id")] int Id,
     [property: JsonPropertyName("name")] string Name,
-    [property: JsonPropertyName("ships")] ushort Ships);
+    [property: JsonPropertyName("ships")] int Ships);
